style: stylize the login page
