+++ frontend/src/MyRoutes.tsx
@@ -6,7 +6,7 @@ import Book from "./pages/Book";
 import SignUp from "./pages/SignUp";
 import FixedNavLayout from "./layouts/FixedNavLayout";
 import BasicLayout from "./layouts/BasicLayout";
-import Login from "./pages/forms/login/Login";
+import Login from "./pages/Login";
 
 const MyRoutes: React.FC = () => {
 	return (
